+++ src/app/_utils/deckProviders/KyberdecksDeckProvider.ts
@@ -1,5 +1,5 @@
 import { DeckSource } from '../deckTypes';
-import { DeckProviderBase } from './DeckProviderBase';
+import { DeckProviderBase } from './core/DeckProviderBase';
 
 export class KyberdecksDeckProvider extends DeckProviderBase {
     public override readonly source = DeckSource.KyberDecks;
