+++ src/app/_utils/deckProviders/SwumetastatsDeckProvider.ts
@@ -1,5 +1,5 @@
 import { DeckSource } from '../deckTypes';
-import { DeckProviderBase } from './DeckProviderBase';
+import { DeckProviderBase } from './core/DeckProviderBase';
 
 export class SwumetastatsDeckProvider extends DeckProviderBase {
     public override readonly source = DeckSource.SWUMetaStats;
